frontend/requirements.txt
Flask
requests
python-dotenv
flask-cors
rich
flasgger
gunicorn
